Replace deprecated TypeScript APIs
