+++ .gitignore
@@ -33,7 +33,8 @@ node_modules
 .LSOverride
 
 # Icon must ends with two \r.
-Icon
+Icon
+
 
 # Thumbnails
 ._*
@@ -42,3 +43,5 @@ Icon
 .Spotlight-V100
 .Trashes
 
+# Output files
+index.js
