+++ package.json
@@ -2,13 +2,6 @@
   "name": "@taskworld/typescript-custom-transformer-l10n",
   "description": "A TypeScript Custom AST transformer that inlines localization strings into the source code",
   "version": "3.0.0-0",
-  "main": "transformer.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "files": [
-    "transformer.js"
-  ],
   "author": "developers@taskworld.com",
   "license": "MIT",
   "repository": {
@@ -19,16 +12,39 @@
     "url": "https://github.com/taskworld/typescript-custom-transformer-l10n/issues"
   },
   "homepage": "https://github.com/taskworld/typescript-custom-transformer-l10n#readme",
+  "main": "index.js",
+  "files": [
+    "index.js"
+  ],
+  "scripts": {
+    "build": "npx tsc transformer.ts --target ES2020 --module CommonJS --skipLibCheck && mv transformer.js index.js",
+    "test": "jest",
+    "preversion": "npm run build",
+    "version": "npm publish",
+    "prepare": "husky install"
+  },
   "devDependencies": {
+    "@jest/globals": "^29.5.0",
+    "husky": "^8.0.3",
     "jest": "^29.5.0",
     "prettier": "^2.8.7",
-    "typescript": "^3.2.2"
+    "ts-jest": "^29.1.0",
+    "typescript": "4.9.5"
   },
   "peerDependencies": {
-    "typescript": "^3.2.2"
+    "typescript": "^4.0.0"
   },
   "resolutions": {
     "**/minimatch": "3.1.2",
     "**/lodash": "4.17.21"
+  },
+  "prettier": {
+    "semi": false,
+    "singleQuote": true,
+    "trailingComma": "es5"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node"
   }
 }
