+++ transformer.test.ts
@@ -1,7 +1,8 @@
-const ts = require('typescript')
-const transformer = require('./transformer')
+import * as ts from 'typescript'
+import transformer from './transformer'
+import { it, expect } from '@jest/globals'
 
-const demoData = {
+const l10nData = {
   en: {
     test: 'Test',
     hello: 'Hello',
@@ -11,20 +12,20 @@ const demoData = {
     hello: 'Nyan',
   },
   'a+1': {
-    x: 1,
+    x: '1',
   },
   'a-1': {
-    x: 2,
+    x: '2',
   },
 }
 
-function test(input, output) {
+function test(input: string, output: string) {
   it(input, () => {
     expect(
       ts
         .transpileModule(input, {
           transformers: {
-            before: [transformer({ l10nData: demoData })],
+            before: [transformer({ l10nData })],
           },
         })
         .outputText.trim()
@@ -48,17 +49,17 @@ test(
 test(
   '__.dual(1, "test", "hello")',
   'var l10n$test = { en: "Test", nyan: "Meow" }, l10n$hello = { en: "Hello", nyan: "Nyan" };' +
-  '__.dual(1, __.$("test", l10n$test), __.$("hello", l10n$hello));'
+    '__.dual(1, __.$("test", l10n$test), __.$("hello", l10n$hello));'
 )
 test(
   '__.dual.string(1, "test", "hello")',
   'var l10n$test = { en: "Test", nyan: "Meow" }, l10n$hello = { en: "Hello", nyan: "Nyan" };' +
-  '__.dual.string(1, __.$("test", l10n$test), __.$("hello", l10n$hello));'
+    '__.dual.string(1, __.$("test", l10n$test), __.$("hello", l10n$hello));'
 )
 test(
   '__("test", { x: __("hello") })',
   'var l10n$test = { en: "Test", nyan: "Meow" }, l10n$hello = { en: "Hello", nyan: "Nyan" };' +
-  '__(__.$("test", l10n$test), { x: __(__.$("hello", l10n$hello)) });'
+    '__(__.$("test", l10n$test), { x: __(__.$("hello", l10n$hello)) });'
 )
 
 // Ensures that the transformer doesn't break when the l10nData has a key that
@@ -66,16 +67,16 @@ test(
 test(
   'let x = () => [__("a+1"), __.string("a-1")]',
   'var l10n$a_1 = {}, l10n$a_1_ = {};' +
-  'var x = function () { return [__(__.$("a+1", l10n$a_1)), __.string(__.$("a-1", l10n$a_1_))]; };'
+    'var x = function () { return [__(__.$("a+1", l10n$a_1)), __.string(__.$("a-1", l10n$a_1_))]; };'
 )
 
 it('runs the example in README', async () => {
   const readme = require('fs').readFileSync('README.md', 'utf8')
   const prettier = require('prettier')
-  let inputCode
-  let expectedCode
+  let inputCode: string = ''
+  let expectedCode: string = ''
   readme.replace(
-    /```js\s*((?:function Tutorial|var l10n)[^]*?)```/g,
+    /```(j|t)s\s*((?:function Tutorial|var l10n)[^]*?)```/g,
     (a, code) => {
       if (!inputCode) inputCode = code.trim()
       else if (!expectedCode) expectedCode = code.trim()
@@ -91,7 +92,7 @@ it('runs the example in README', async () => {
         },
         compilerOptions: {
           charset: 'utf8',
-          jsx: 'preserve',
+          jsx: ts.JsxEmit.Preserve,
         },
       }).outputText,
       { ...prettierConfig, parser: 'typescript' }
